# WIP: Integrate md5-simd