# ABANDONED: Slower.

## Description

Biggest change is that `hash.Reader` now implements `io.Closer` and must be closed to release resources.

Chaining is also now allowed and the readers are merged.

However, testing with warp shows significant performance regressions:

Using `warp put --md5 -duration=1m -concurrent=16` against 4 XL servers, 1 drive each.

```
λ warp cmp warp-put-10MB-DISABLED.csv.zst warp-put-10MB-ENABLED2.csv.zst
-------------------
Operation: PUT
Operations: 2694 -> 2492
* Average: -7.40% (-33.1 MiB/s) throughput, -7.40% (-3.3) obj/s
* Fastest: -15.72% (-88.4 MiB/s) throughput, -15.72% (-8.8) obj/s
* 50% Median: -5.77% (-25.6 MiB/s) throughput, -5.77% (-2.6) obj/s
* Slowest: +102.12% (+147.6 MiB/s) throughput, +102.12% (+14.8) obj/s

λ warp cmp warp-put-10MB-DISABLED2.csv.zst warp-put-10MB-ENABLED3.csv.zst
-------------------
Operation: PUT
Operations: 2625 -> 2282
* Average: -13.09% (-57.1 MiB/s) throughput, -13.09% (-5.7) obj/s
* Fastest: -22.22% (-122.0 MiB/s) throughput, -22.22% (-12.2) obj/s
* 50% Median: -12.42% (-54.9 MiB/s) throughput, -12.42% (-5.5) obj/s
* Slowest: +91.84% (+124.6 MiB/s) throughput, +91.84% (+12.5) obj/s
```
However my system is very close to being IO limited.

Running in FS mode paints a clearer picture, 2 runs:

```
λ warp cmp warp-put-FS-DISABLED.csv.zst warp-put-FS-ENABLED.csv.zst
-------------------
Operation: PUT
Operations: 4616 -> 2336
Duration: 1m0s -> 59s
* Average: -49.47% (-380.1 MiB/s) throughput, -49.47% (-38.0) obj/s
* Fastest: -25.56% (-237.3 MiB/s) throughput, -25.56% (-23.7) obj/s
* 50% Median: -55.34% (-427.7 MiB/s) throughput, -55.34% (-42.8) obj/s
* Slowest: -59.37% (-389.8 MiB/s) throughput, -59.37% (-39.0) obj/s

λ warp cmp warp-put-FS-DISABLED2.csv.zst warp-put-FS-ENABLED2.csv.zst
-------------------
Operation: PUT
Operations: 4513 -> 2279
Duration: 1m0s -> 59s
* Average: -49.59% (-372.7 MiB/s) throughput, -49.59% (-37.3) obj/s
* Fastest: -27.42% (-253.8 MiB/s) throughput, -27.42% (-25.4) obj/s
* 50% Median: -54.30% (-409.0 MiB/s) throughput, -54.30% (-40.9) obj/s
* Slowest: -56.70% (-339.3 MiB/s) throughput, -56.70% (-33.9) obj/s
```

With `--concurrent=32`:

```
λ warp cmp warp-put-FS-32C-DISABLED.csv.zst warp-put-FS-32C-ENABLED.csv.zst
-------------------
Operation: PUT
Operations: 4827 -> 3149
* Average: -34.68% (-278.0 MiB/s) throughput, -34.68% (-27.8) obj/s
* Fastest: -21.96% (-211.5 MiB/s) throughput, -21.96% (-21.1) obj/s
* 50% Median: -37.09% (-299.9 MiB/s) throughput, -37.09% (-30.0) obj/s
* Slowest: -43.06% (-276.0 MiB/s) throughput, -43.06% (-27.6) obj/s
```
## How to test this PR?

Upload stuff with MD5.

## Types of changes
- [x] New feature (non-breaking change which adds functionality)