@minio-trusted 2020-05-07
```
Error running mint automation
+ mkdir -p 9518-4b83269/gopath/src/github.com/minio
+ git -C 9518-4b83269/gopath/src/github.com/minio clone --quiet https://github.com/minio/minio.git
+ git -C 9518-4b83269/gopath/src/github.com/minio/minio remote add minio https://github.com/minio/minio.git
+ git -C 9518-4b83269/gopath/src/github.com/minio/minio fetch --quiet minio pull/9518/head:pr9518
+ git -C 9518-4b83269/gopath/src/github.com/minio/minio checkout --quiet pr9518
+ GO111MODULE=on
+ GOPATH=/home/mint-auto/mint-auto/9518-4b83269/gopath
+ make -C 9518-4b83269/gopath/src/github.com/minio/minio --quiet
go: errors parsing go.mod:
/home/mint-auto/mint-auto/9518-4b83269/gopath/src/github.com/minio/minio/go.mod:133: replacement directory appears to be Windows path (on a non-windows system)
go: errors parsing go.mod:
/home/mint-auto/mint-auto/9518-4b83269/gopath/src/github.com/minio/minio/go.mod:133: replacement directory appears to be Windows path (on a non-windows system)
make: *** [build] Error 1

```